moved the Extension of RTCAudioSession to a different file.
